Refactor code - add module classes for the different components
