+++ src/assets/styles/index.scss
@@ -63,6 +63,6 @@ a:hover {
   color: rgb(255, 94, 0);
 }
 
-.search-btn {
-  // border: rgb(255, 255, 255) solid;
+.centered-content {
+  margin: 10rem auto;
 }
